Fix `Trakeva_sqlite.iterator` (read-write access)
There was a “bug”  while iterating on a collection and modifying it: Sqlite was
looping infinitely (returning the new value or an already visited key-value
pair).

This patch adds tests that show the issue and fixes it. The new implementation
of the iterator uses 2 different queries: one to get all the distinct keys, and
one per key to actually get and return the value.