Typo 'falg' -> 'flag'
The short circuit of ||, ( I guess) does not trigger this when you do a `make configure`. But I somehow got this error when installing trakeva via opam. Although that was on 4.02, before I downgraded and the problem went away. I'm not going to untangle all the build dependencies of this, just noting that something weird happened.